grana.text: now based on elk.graph.text

+++ plugins/de.cau.cs.kieler.kiml.grana.text/src-gen/de/cau/cs/kieler/kiml/grana/text/grana/FloatRange.java
@@ -24,7 +24,7 @@ public interface FloatRange extends Range
 {
   /**
    * Returns the value of the '<em><b>Values</b></em>' attribute list.
-   * The list contents are of type {@link java.lang.Float}.
+   * The list contents are of type {@link java.lang.Double}.
    * <!-- begin-user-doc -->
    * <p>
    * If the meaning of the '<em>Values</em>' attribute list isn't clear,
@@ -36,6 +36,6 @@ public interface FloatRange extends Range
    * @model unique="false"
    * @generated
    */
-  EList<Float> getValues();
+  EList<Double> getValues();
 
 } // FloatRange
